requirements.txt
psycopg2
fastapi
pydantic
sqlalchemy
uvicorn
python-dotenv